tools: filter browser API to call according to mcp method
